Fetching Single Result
The social media app can now enter a single post, but not yet read comments nor comment on the post.

The post card displays the user who posted it. Avatar and name have been added to the head of the post card.
Clicking the avatar or name sends the viewer to the persons profile, but the back button in profileDetail is not optimal.